backend/requirements.txt
# ─────────────────────────────────────────
# Web Framework
# ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2

# ─────────────────────────────────────────
# Database
# ─────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9     # Sync driver required by Alembic migrations
alembic==1.13.3
aiosqlite==0.20.0          # SQLite async driver for local mode

# ─────────────────────────────────────────
# Task Queue
# ─────────────────────────────────────────
celery==5.4.0
redis==5.1.1
flower==2.0.1

# ─────────────────────────────────────────
# File Storage (MinIO / S3)
# ─────────────────────────────────────────
boto3==1.35.0
minio==7.2.9

# ─────────────────────────────────────────
# Data validation
# ─────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0

# ─────────────────────────────────────────
# AI / ML (lightweight — heavy models in OCR service)
# ─────────────────────────────────────────
sentence-transformers==3.2.1
ollama==0.3.3               # Ollama Python client
groq==0.11.0                # Groq cloud API client
sympy==1.13.3               # Math verification

# ─────────────────────────────────────────
# Export
# ─────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5
weasyprint==62.3            # PDF generation
gspread==6.1.2              # Google Sheets
google-auth==2.36.0

# ─────────────────────────────────────────
# Email
# ─────────────────────────────────────────
resend==2.2.0

# ─────────────────────────────────────────
# Utils
# ─────────────────────────────────────────
pillow==10.4.0
python-dotenv==1.0.1
structlog==24.4.0           # Structured logging
tenacity==9.0.0             # Retry logic
qrcode[pil]==8.0
shortuuid==1.0.13           # Short unique IDs for share tokens
